Corrected some bugs in stats. There's still one.
Some costumer counts are transformed into provider counts incorrectly.
As an example in 11.txt when orig = 2, 3 is found as a costumer by 2 but then 3 is found as a provider by 4 and it's route type is transformed from costumer to provider.

The stats report 14 provider routes and 6 costumer routes. This should be 11, 9. 14-3 = 11 and 6+3 = 9. So 3 paths are wrongly morphed.